refactor(notes): rename service files
