+++ pkg/notes/adaptor/controller/note.ts
@@ -5,8 +5,8 @@ import type { AccountID } from '../../../accounts/model/account.js';
 import type { ID } from '../../../id/type.js';
 import type { AccountModule } from '../../../intermodule/account.js';
 import type { NoteVisibility } from '../../model/note.js';
-import type { CreateNoteService } from '../../service/create.js';
-import type { FetchNoteService } from '../../service/fetch.js';
+import type { CreateNoteService } from '../../service/createNote.js';
+import type { FetchNoteService } from '../../service/fetchNote.js';
 import {
   type CreateNoteResponseSchema,
   type GetNoteResponseSchema,
