+++ pkg/notes/service/fetchNote.ts
